Change wallbox id from uint8 to uint32
This way the Brick can use the UID. We also save the filename
as Base58 now to reduce the file size and make it human readable.

+++ software/src/communication.h
@@ -243,7 +243,7 @@ typedef struct {
 
 typedef struct {
 	TFPMessageHeader header;
-	uint8_t wallbox_id;
+	uint32_t wallbox_id;
 	uint8_t year;
 	uint8_t month;
 	uint8_t day;
@@ -260,7 +260,7 @@ typedef struct {
 
 typedef struct {
 	TFPMessageHeader header;
-	uint8_t wallbox_id;
+	uint32_t wallbox_id;
 	uint8_t year;
 	uint8_t month;
 	uint8_t day;
@@ -276,7 +276,7 @@ typedef struct {
 
 typedef struct {
 	TFPMessageHeader header;
-	uint8_t wallbox_id;
+	uint32_t wallbox_id;
 	uint8_t year;
 	uint8_t month;
 	uint8_t day;
@@ -290,7 +290,7 @@ typedef struct {
 
 typedef struct {
 	TFPMessageHeader header;
-	uint8_t wallbox_id;
+	uint32_t wallbox_id;
 	uint8_t year;
 	uint8_t month;
 	uint8_t day;
